Fix MSVC crash for strutil's replace() As it turns out, MSVC stores std::string::npos as uint64, which is not unsigned long.

+++ src/common/strutil.cpp
@@ -66,7 +66,7 @@ std::string replace(const std::string &str, const std::string &what, const std::
 		CreateException("replace(): replacing {} with {} may create an infinite loop", what, with);
 	
 	std::string result = str;
-	unsigned long int position = std::string::npos;
+	unsigned long long position = std::string::npos;
 
 	while ((position = str.find(what)) != std::string::npos) 
 		result.replace(position, what.length(), with);
